% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidPlot.R
\name{lipidPlot}
\alias{lipidPlot}
\title{plots lipidonomic data based on specifications set by parse_raw and summarise_lipids functions}
\usage{
lipidPlot(object)
}
\arguments{
\item{object}{takes an object generated by summarise_lipids function}
}
\description{
This combines head with a number of columns
}
