% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw.R
\docType{class}
\name{lipids-class}
\alias{lipids-class}
\title{parsing the raw experimental data}
\arguments{
\item{filename}{tsv filename}

\item{lipid_class}{filter data by lipid class e.g. FFA or LPL}

\item{selected_chains}{filter data by selected_chains e.g. 10:0 etc}

\item{treat_groups}{filter data by treatment groups}

\item{columns}{declare columns which contain experimental subject data e.g. M1, M2, etc for mouse 1 etc.}

\item{filter_method}{apply outlier removal e.g. mf+replacemedian for median filtering and median replacment by default.}

\item{filter_cutoff}{outlier removal cutoff which is 2 by default}
}
\description{
This combines head with a number of columns
}
