% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_lipids.R
\docType{class}
\name{lipidSummary-class}
\alias{lipidSummary-class}
\title{parsing the raw experimental data}
\arguments{
\item{object}{takes an object generated by parse_raw function}

\item{grouping}{declare the grouping variable to summarise by e.g. CHAIN, SATURATION, REGION, etc}

\item{statistic}{choose summary statistic e.g. 'means +/- se', 'response' for difference between treatments, and 'percentage' for percentage out of total concentration}

\item{region_order}{declare region order}
}
\description{
This combines head with a number of columns
}
