% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_output.R
\name{plotting_output}
\alias{plotting_output}
\title{plots a combination plot e.g. CHAIN + SATURATION. Replacement for lipidPlot}
\usage{
plotting_output(filename, treat_groups = NULL, columns,
  select_chains = NULL, statistic = "means +/- se")
}
\arguments{
\item{treat_groups}{takes an object generated by summarise_lipids function}

\item{columns}{takes an object generated by summarise_lipids function}

\item{select_chains}{takes an object generated by summarise_lipids function}

\item{statistic}{takes an object generated by summarise_lipids function}
}
\description{
This makes a combination plot
}
