% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reject_outliers.R
\name{reject_outliers}
\alias{reject_outliers}
\title{saturation look up}
\usage{
reject_outliers(df, columns, cutoff = 2, method = "mf+replacemedian")
}
\arguments{
\item{df}{enter data frame}

\item{columns}{key by which to lookup e.g. CHAIN or SATURATION}

\item{cutoff}{key by which to lookup e.g. CHAIN or SATURATION}

\item{method}{key by which to lookup e.g. CHAIN or SATURATION}
}
\description{
Look up the saturation value / colour / order
}
