% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_lookup.R
\name{sat_lookup}
\alias{sat_lookup}
\title{saturation look up}
\usage{
sat_lookup(df, keys = "CHAIN")
}
\arguments{
\item{df}{enter data frame}

\item{keys}{key by which to lookup e.g. CHAIN or SATURATION}
}
\description{
Look up the saturation value / colour / order
}
